fix(vault): remove unnecessary cast causing clippy error

+++ contracts/vault/src/lib.rs
@@ -4716,7 +4716,8 @@ impl VaultDAO {
         // Exponential backoff: initial_backoff << (retry_count - 1), capped at 7 days (120,960 ledgers)
         let max_backoff = 17_280 * 7; // 7 days in ledgers
         let exponent = core::cmp::min(retry_state.retry_count - 1, 30); // Prevent overflow
-        let backoff = (retry_config.initial_backoff_ledgers as u64)
+        let backoff = retry_config
+            .initial_backoff_ledgers
             .checked_shl(exponent as u32)
             .unwrap_or(max_backoff)
             .min(max_backoff);
